# Fix critical data race conditions + improve performance