--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/edwarnicke/serialize
 
 go 1.13
 
-require go.uber.org/goleak v1.1.10
+require (
+	github.com/stretchr/testify v1.4.0
+	go.uber.org/goleak v1.1.10
+)