https://github.com/edwarnicke/serialize/issues/5