@denis-tingajkin 2020-10-17
## Before

```
BenchmarkExecutorAsync
BenchmarkExecutorAsync-8   	 5511322	       217 ns/op

BenchmarkExecutorSync
BenchmarkExecutorSync-8   	 1571877	       742 ns/op
```

## After

```
BenchmarkExecutorAsync
BenchmarkExecutorAsync-8   	 5717280	       205 ns/op


BenchmarkExecutorSync
BenchmarkExecutorSync-8   	 1767159	       667 ns/op
```


## Delta

```
BenchmarkExecutorAsync-16    	        	       -12  ns/op (5.5% faster)
BenchmarkExecutorSync-16     	        	       -75 ns/op (10.10% faster)
```